Refactor: 캐러셀 기능 hooks로  분리
